Improve azure ad sample
