+++ README.md
@@ -1,6 +1,7 @@
 # Microsoft Azure Java SDK for Media Services Samples
-This repository contains the following samples showing how to use the [Azure Java SDK for Media Service](https://github.com/Azure/azure-sdk-for-java/tree/master/services/azure-media).
+This repository contains the following samples showing how to use the [Azure Java SDK for Media Service](https://github.com/Azure/azure-sdk-for-java/tree/0.9/services/azure-media).
 
+* **azure-media-azuread-authentication**: This sample consists in three a console Java applications that show how to use User/Password Authentication, Service Principal SymmetricKey Authentication, and Service Principal Certificate authentication.
 * **azure-media-dynamic-encryption-aes**: This sample is a console Java application that shows how to use AES-128 Dynamic Encryption and Key Delivery Service. It is based on the .NET sample explained in this ACOM article: [https://azure.microsoft.com/en-us/documentation/articles/media-services-protect-with-aes128/](https://azure.microsoft.com/en-us/documentation/articles/media-services-protect-with-aes128/).
 * **azure-media-dynamic-encryption-playready**: This sample is a console Java application that shows how to use PlayReady Dynamic Encryption and License Delivery Service. It is based on the .NET sample explained in this ACOM article: [https://azure.microsoft.com/en-us/documentation/articles/media-services-protect-with-drm/](https://azure.microsoft.com/en-us/documentation/articles/media-services-protect-with-drm/).
 * **azure-media-dynamic-encryption-playreadywidevine**: This sample is a console Java application that shows how to use PlayReady and Widevine (DRM) Dynamic Encryption and License Delivery Service. It is based on the .NET sample explained in this ACOM article: [https://azure.microsoft.com/en-us/documentation/articles/media-services-protect-with-drm/](https://azure.microsoft.com/en-us/documentation/articles/media-services-protect-with-drm/).
@@ -8,12 +9,12 @@ This repository contains the following samples showing how to use the [Azure Jav
 * **azure-media-scale-workflow**: This sample is a console Java application that shows how to scale Encoding Reserved Units and Streaming Endpoints.
 * **azure-media-analytics-indexer**: This sample is a console Java application that shows how to use Azure Media Services to run a media analytics job with the _Azure Media Indexer_ processor. It is based on the .NET sample explained in this ACOM article: [https://azure.microsoft.com/en-us/documentation/articles/media-services-index-content/](https://azure.microsoft.com/en-us/documentation/articles/media-services-index-content/).
 
-These samples rely on the [azure-media maven package v0.9.4](http://mvnrepository.com/artifact/com.microsoft.azure/azure-media/0.9.4).
+These samples rely on the [azure-media maven package v0.9.8](http://mvnrepository.com/artifact/com.microsoft.azure/azure-media/0.9.8).
 
 ```xml
 <dependency>
   <groupId>com.microsoft.azure</groupId>
   <artifactId>azure-media</artifactId>
-  <version>0.9.4</version>
+  <version>0.9.8</version>
 </dependency>
 ```
